--- a/src/univariate/discrete/poissonbinomial.jl
+++ b/src/univariate/discrete/poissonbinomial.jl
@@ -135,6 +135,48 @@ function poissonbinomial_pdf(p::AbstractArray{T,1}) where {T <: Real}
   return S
 end
 
+# Computes the pdf of a poisson-binomial random variable using
+# fast fourier transform
+#
+#     Hong, Y. (2013).
+#     On computing the distribution function for the Poisson binomial
+#     distribution. Computational Statistics and Data Analysis, 59, 41–51.
+#
+function poissonbinomial_pdf_fft(p::AbstractArray{T}) where {T <: Real}
+    n = length(p)
+    ω = 2 * one(T) / (n + 1)
+
+    x = Vector{Complex{T}}(undef, n+1)
+    lmax = ceil(Int, n/2)
+    x[1] = one(T)/(n + 1)
+    for l=1:lmax
+        logz = zero(T)
+        argz = zero(T)
+        for j=1:n
+            zjl = 1 - p[j] + p[j] * cospi(ω*l) + im * p[j] * sinpi(ω * l)
+            logz += log(abs(zjl))
+            argz += atan(imag(zjl), real(zjl))
+        end
+        dl = exp(logz)
+        x[l + 1] = dl * cos(argz) / (n + 1) + dl * sin(argz) * im / (n + 1)
+        if n + 1 - l > l
+            x[n + 1 - l + 1] = conj(x[l + 1])
+        end
+    end
+    [max(0, real(xi)) for xi in _dft(x)]
+end
+
+# A simple implementation of a DFT to avoid introducing a dependency
+# on an external FFT package just for this one distribution
+function _dft(x::Vector{T}) where T
+    n = length(x)
+    y = zeros(complex(float(T)), n)
+    @inbounds for j = 0:n-1, k = 0:n-1
+        y[k+1] += x[j+1] * cis(-π * float(T)(2 * mod(j * k, n)) / n)
+    end
+    return y
+end
+
 #### Sampling
 
 sampler(d::PoissonBinomial) = PoissBinAliasSampler(d)