--- a/test/poissonbinomial.jl
+++ b/test/poissonbinomial.jl
@@ -1,6 +1,37 @@
 using Distributions
 using Test
 
+function naive_esf(x::AbstractVector{T}) where T <: Real
+    n = length(x)
+    S = zeros(T, n+1)
+    states = hcat(reverse.(digits.(0:2^n-1,base=2,pad=n))...)'
+
+    r_states = vec(mapslices(sum, states, dims=2))
+
+    for r in 0:n
+        idx = findall(r_states .== r)
+        S[r+1] = sum(mapslices(x->prod(x[x .!= 0]), states[idx, :] .* x', dims=2))
+    end
+    return S
+end
+
+function naive_pb(p::AbstractVector{T}) where T <: Real
+    x = p ./ (1 .- p)
+    naive_esf(x) * prod(1 ./ (1 .+ x))
+end
+
+# test PoissonBinomial PMF algorithms
+x = [3.5118, .6219, .2905, .8450, 1.8648]
+n = length(x)
+p = x ./ (1 .+ x)
+
+naive_sol = naive_pb(p)
+
+@test Distributions.poissonbinomial_pdf_fft(p) ≈ naive_sol
+@test Distributions.poissonbinomial_pdf(p) ≈ naive_sol
+
+@test Distributions.poissonbinomial_pdf_fft(p) ≈ Distributions.poissonbinomial_pdf(p)
+
 # Test the special base where PoissonBinomial distribution reduces
 # to Binomial distribution
 for (p, n) in [(0.8, 6), (0.5, 10), (0.04, 20)]