@johnczito 2020-10-16
Thanks! Haven't looked at this in detail yet, but I'm in favor of keeping the old implementation around so you can write some unit tests comparing the two.