@bdeonovic 2020-10-16
No problem, probably tomorrow. its a nice day when a simple for loop gives 50x speedup to a function :)