@johnczito 2020-10-16
Just check that `poissonbinomial_pdf` and `poissonbinomial_pdf_fft` give approximately the same thing.