@bdeonovic 2020-10-16
Added some additional tests, but I should point out that my PR had alread passed the tests that were implemented before for poisson binomial