@bdeonovic 2020-10-20
@johnczito The biscarri method requires a solid working fft filter function. In my package I imported the corresponding functions from the DSP package. I don't get the same speed up with my own written fft filter. 

Below  esf_sum is the recursive algorithm implemented in this PR, esf_dc_group is the biscarri method with my own fft filt code, and esf_dc_fft is the biscarri method with the fft filt from DSP package. I think using the simple recursive algorithm I have already provided is enough as it won't require any more package dependencies and the speed up is not that significant. I propose to go ahead and merge. 

```julia
julia> x = exp.(randn(15000))
julia> @benchmark esf_dc_group(x, 4)
BenchmarkTools.Trial: 
  memory estimate:  1.95 MiB
  allocs estimate:  9
  --------------
  minimum time:     175.359 ms (0.00% GC)
  median time:      178.497 ms (0.00% GC)
  mean time:        181.368 ms (0.07% GC)
  maximum time:     200.614 ms (0.00% GC)
  --------------
  samples:          28
  evals/sample:     1

julia> @benchmark esf_sum(x)
BenchmarkTools.Trial: 
  memory estimate:  117.33 KiB
  allocs estimate:  2
  --------------
  minimum time:     44.550 ms (0.00% GC)
  median time:      44.788 ms (0.00% GC)
  mean time:        45.534 ms (0.00% GC)
  maximum time:     53.893 ms (0.00% GC)
  --------------
  samples:          110
  evals/sample:     1

julia> @benchmark esf_dc_fft(x,4)
BenchmarkTools.Trial: 
  memory estimate:  2.41 MiB
  allocs estimate:  60
  --------------
  minimum time:     31.536 ms (0.00% GC)
  median time:      32.574 ms (0.00% GC)
  mean time:        33.238 ms (0.61% GC)
  maximum time:     44.119 ms (11.66% GC)
  --------------
  samples:          151
  evals/sample:     1
```